refactor(fuse): group recluster implementation
