+++ src/query/storages/fuse/src/operations/mod.rs
@@ -55,7 +55,12 @@ pub use read::DeserializeDataTransform;
 pub use read::ReadState;
 pub use read::need_reserve_block_info;
 pub use read::row_fetch_processor;
+pub use recluster::CandidateScore;
+pub use recluster::ReclusterCandidateWindow;
+pub use recluster::ReclusterFinalCarry;
 pub use recluster::ReclusterMode;
+pub use recluster::ReclusterMutator;
+pub use recluster::SelectedReclusterSegment;
 pub use replace_into::*;
 pub use snapshot_hint::*;
 pub use table_index::do_refresh_table_index;
